requirements.txt
python-dotenv
wheel
langchain
langchain_community
unstructured
langchain-unstructured
unstructured-client
unstructured[pdf]
python-magic
langchain-google-genai
pypdf
beautifulsoup4
Arxiv
pymupdf
Wikipedia
youtube-transcript-api
pytube
langchain_core
assemblyai
streamlit
langchain_huggingface
fpdf